Remove hardcoded poller instances
The cron poller-wrapper is set to a hardcoded "2" poller instances. This is deprecated usage. Poller instances are fully configurable via the UI now.
I have recently added quite a few more hosts to observium and wondered why my poller was taking so long to complete! Removing the "2" has allowed the poller-wrapper to create more instances and resolved my issues.